+++ src/main/java/de/ait_tr/g_40_shop/security/sec_service/AuthService.java
@@ -34,7 +34,7 @@ public class AuthService {
         if (passwordEncoder.matches(inboundUser.getPassword(), foundUser.getPassword()))
         {
             if (!foundUser.isActive()){
-                return new TokenResponseDto("your account", "is inactive");
+                throw  new RuntimeException("your account is inactive");
             }
             String accessToken = tokenService.generateAccessToken(foundUser);
             String refreshToken = tokenService.generateRefreshToken(foundUser);
