+++ src/main/java/de/ait_tr/g_40_shop/service/Interfaces/UserService.java
@@ -6,4 +6,5 @@ import org.springframework.security.core.userdetails.UserDetailsService;
 public interface UserService extends UserDetailsService {
 
     void register(User user);
+    void confirmAccount(String code);
 }
